Refactor map creation
This lays the ground to have per-dataset maps on the `/datasets/` routes.
