+++ src/pages/map.astro
@@ -19,67 +19,3 @@ import "@styles/map-styles.css";
   </section>
   <script src="@scripts/main-map.mjs"></script>
 </Layout>
-
-<style>
-  .auto-search-wrapper {
-    display: none;
-    left: calc(34px + 20px);
-    position: absolute;
-    top: 10px;
-    width: min(250px, calc(100vw - 34px - 20px - 48px - 20px));
-    z-index: 999;
-
-    input {
-      background-clip: padding-box;
-      border: 2px solid rgba(0, 0, 0, 0.2);
-      border-radius: 5px;
-      padding: 10px 45px 10px 35px;
-    }
-
-    :global(.auto-results-wrapper.auto-is-active) {
-      background-clip: padding-box;
-      border-radius: 5px;
-      border: 2px solid rgba(0, 0, 0, 0.2);
-      font-size: 18px;
-      max-height: calc(100dvh - 200px - 5dvh - 10px);
-      max-width: calc(100vw - 34px - 20px - 10px);
-      overflow-y: auto;
-      transform: translateY(2px);
-      width: max-content;
-    }
-
-    :global(ul) {
-      min-width: min(500px, 80vw);
-    }
-
-    :global(li) {
-      outline: 1px solid rgba(0 0 0 / 0.05);
-
-      :global(span) {
-        float: right;
-      }
-    }
-
-    :global(.auto-selected) {
-      background-color: var(--theme-surface-1);
-    }
-
-    :global(mark) {
-      background: none;
-      color: var(--theme-primary);
-      font-weight: bold;
-      outline: 1px solid rgba(218 165 32 / 0.6);
-    }
-  }
-
-  :global(.pill) {
-    color: white;
-    background: var(--theme-surface-2);
-    border-radius: var(--theme-button-border-radius);
-    display: inline-block;
-    font-size: 12px;
-    font-weight: bold;
-    margin-bottom: 8px;
-    padding: 2px 8px;
-  }
-</style>
